Make tmux session save safer

+++ scripts/tmux_session_auto_save_quit.sh
@@ -1,4 +1,13 @@
 #!/usr/bin/env bash
 
-tmux run-shell '~/.tmux/plugins/tmux-resurrect/scripts/save.sh'
+# `^bash$` matches this script running
+commands=$(tmux list-panes -a -F "#{pane_current_command}" | grep --extended-regexp --invert-match "^zsh$|^bash$")
+status=$?
+if [[ $status -eq 0 ]]; then
+  commands=${commands//$'\n'/ }
+  echo "Panes found running commands: $commands" >&2
+  exit 1
+fi
+
+tmux run-shell "$HOME/.tmux/plugins/tmux-resurrect/scripts/save.sh"
 tmux kill-server
